Fixes bugs with console display on at least 2.x OSes.
The original method appears to be an attempt to display all characters in the
buffer even if the buffer contains null characters. However, the
implementation was buggy and occasionally printed characters past the
original buffer due to the lack of null termination if a full buffer was
read.

Also, the iRecovery prompt is changed to "iRecovery>" from "] " so that it
and iBoot's output prompt "] " may be distinguished.